requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Models
timm>=0.9.12
transformers>=4.38.0

# Data / augmentation
albumentations>=1.3.1
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# YOLO (optional — for leaf detection)
ultralytics>=8.0.0

# Evaluation / visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
